separating  the OrderView components into separate files
Temporaly, the components can be moved to a differently directory than the one the components are in now
